Issue #6974 - improve the WebSocketBufferPoolTest
